ProgramExecutor: Implement a program pause dialog
It's probably a good idea to clearly signal the operator that the
program has been paused by a pause step. The program execution is
resumed either by pressing the Continue button or the enter key. Pauses
are likely to be used for implementing critical manual timings so highly
visible indicator and easy to press continue key are required.